Make DBManager abstract and implement MySQLController
